ai-unit-test-generator/python/pyproject.toml
[project]
name = "ai-unit-test-generator-backend"
version = "0.1.0"
description = "Python backend for AI Unit Test Generator VS Code extension"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "langsmith>=0.1.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
openai = ["langchain-openai>=0.1.0"]
ollama = ["langchain-ollama>=0.1.0"]
all = [
    "langchain-openai>=0.1.0",
    "langchain-ollama>=0.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "langchain-openai>=0.1.0",
    "langchain-ollama>=0.1.0",
]
